Moved classes to appropriate package
