+++ src/main/groovy/org/shipkit/gradle/notes/FetchReleaseNotesTask.java
@@ -1,10 +1,11 @@
-package org.shipkit.gradle;
+package org.shipkit.gradle.notes;
 
 import org.gradle.api.DefaultTask;
 import org.gradle.api.tasks.Input;
 import org.gradle.api.tasks.Optional;
 import org.gradle.api.tasks.OutputFile;
 import org.gradle.api.tasks.TaskAction;
+import org.shipkit.gradle.ReleaseConfiguration;
 import org.shipkit.internal.gradle.util.ReleaseNotesSerializer;
 import org.shipkit.internal.notes.generator.ReleaseNotesGenerator;
 import org.shipkit.internal.notes.generator.ReleaseNotesGenerators;
