+++ src/main/groovy/org/shipkit/internal/gradle/ReleaseNotesPlugin.java
@@ -4,8 +4,8 @@ import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.shipkit.gradle.ReleaseConfiguration;
-import org.shipkit.gradle.FetchReleaseNotesTask;
-import org.shipkit.gradle.UpdateReleaseNotesTask;
+import org.shipkit.gradle.notes.FetchReleaseNotesTask;
+import org.shipkit.gradle.notes.UpdateReleaseNotesTask;
 import org.shipkit.internal.gradle.configuration.ReleaseConfigurationPlugin;
 import org.shipkit.gradle.notes.FetchContributorsTask;
 import org.shipkit.internal.gradle.contributors.ContributorsPlugin;
