+++ src/main/groovy/org/shipkit/internal/gradle/release/BintrayReleasePlugin.java
@@ -5,7 +5,7 @@ import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
-import org.shipkit.gradle.UpdateReleaseNotesTask;
+import org.shipkit.gradle.notes.UpdateReleaseNotesTask;
 import org.shipkit.internal.gradle.ShipkitBintrayPlugin;
 import org.shipkit.internal.gradle.ReleaseNotesPlugin;
 import org.shipkit.internal.gradle.git.GitPlugin;
