+++ src/test/groovy/org/shipkit/gradle/notes/UpdateReleaseNotesTaskTest.groovy
@@ -1,9 +1,10 @@
-package org.shipkit.gradle
+package org.shipkit.gradle.notes
 
 import org.gradle.api.GradleException
 import org.gradle.testfixtures.ProjectBuilder
 import org.junit.Rule
 import org.junit.rules.TemporaryFolder
+import org.shipkit.gradle.notes.UpdateReleaseNotesTask
 import spock.lang.Specification
 
 class UpdateReleaseNotesTaskTest extends Specification {
